arm: Allow disabling PIE for dynamically linked executables
PIE can and does break some binaries that were built without it,
resulting in relocation-related issues (addressing memory
wrongly assumed to be at a fixed position).

Use this as a last resort measure for compatibility with pre-JB
binary blobs, and only if confirmed to fix a broken executable.
If you're not sure, do NOT enable this on a device.

Conflicts:

	core/combo/TARGET_linux-arm.mk

Change-Id: I3b7001e682a6f7a845e62278fe2afa9ccf5552a4